separate minion reduction rules
