+++ conjure_oxide/src/rules/mod.rs
@@ -1,2 +1,3 @@
 mod base;
 mod cnf;
+mod minion;
